Move logging interceptor to new module
